refactor: major refactorings and folder restructure
