+++ src/data_store/implementations/bytes/bytes.rs
@@ -1,6 +1,6 @@
 use crate::data_store::errors::ValueError;
 
-use super::super::{store::KeyValueStore, value_entry::ValueEntry};
+use crate::data_store::{store::KeyValueStore, value_entry::ValueEntry};
 use std::time::{Duration, Instant};
 
 impl KeyValueStore {
