+++ src/data_store/implementations/bytes/mod.rs
@@ -1,0 +1,2 @@
+mod bytes;
+mod tests;
