+++ src/data_store/implementations/hsets/hsets.rs
@@ -1,6 +1,6 @@
 use crate::data_store::errors::ValueError;
 
-use super::super::{store::KeyValueStore, value_entry::ValueEntry};
+use crate::data_store::{store::KeyValueStore, value_entry::ValueEntry};
 use std::{
     collections::HashSet,
     time::{Duration, Instant},
@@ -170,3 +170,5 @@ impl KeyValueStore {
         }
     }
 }
+
+pub mod hsets_tests {}
