+++ src/data_store/implementations/integers/mod.rs
@@ -1,0 +1,2 @@
+mod integers;
+mod tests;
