+++ src/data_store/implementations/lists/lists.rs
@@ -1,6 +1,6 @@
 use crate::data_store::errors::ValueError;
 
-use super::super::{store::KeyValueStore, value_entry::ValueEntry};
+use crate::data_store::{store::KeyValueStore, value_entry::ValueEntry};
 use std::{
     collections::VecDeque,
     time::{Duration, Instant},
