+++ src/data_store/mod.rs
@@ -1,5 +1,6 @@
 mod errors;
 mod implementations;
 pub mod store;
+mod tests;
 mod types;
 mod value_entry;
